backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-multipart==0.0.17
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
aiofiles==24.1.0
email-validator==2.2.0
google-auth==2.37.0
requests==2.32.3
